Add Dockerfile and update README
Added Dockerfile to create a functional Docker environment for Ubuntu 20.04 with necessary packages for Linux practice. It contains libraries to emulate OS behavior and build essentials. Updated Readme.md to guide users to build and run the Docker image. This setup provides users with a quick and easy environment to study and experiment with Linux.